requirements.txt
# Core Testing Framework
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
allure-pytest==2.13.5

# Web Automation
selenium==4.27.0
webdriver-manager==4.0.2

# API Testing
requests==2.32.3
requests-toolbelt==1.0.0

# Data Processing & Reporting
pandas==2.2.3
openpyxl==3.1.5

# Configuration & Environment
python-dotenv==1.0.1
PyYAML==6.0.2

# Database
mysql-connector-python==8.3.0

# Utilities
colorama==0.4.6
